tools/character-generator/pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "character-generator"
version = "0.1.0"
description = "Batch-generate Dragonbane character portraits with FLUX, face detection, and background removal."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "accelerate>=1.10.0",
  "diffusers>=0.35.0",
  "huggingface-hub>=0.34.0",
  "numpy>=2.0.0",
  "pillow>=11.0.0",
  "pyyaml>=6.0.2",
  "torch>=2.8.0",
  "torchvision>=0.23.0",
  "transformers>=4.56.0",
  "ultralytics>=8.3.0",
]

[project.optional-dependencies]
test = ["pytest>=8.3.0"]

[project.scripts]
character-generator = "character_generator.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
